Ensure websockets actually disconnect immediately when a server is suspended
When a server status switches to suspended, immediately close any open websocket connections. Additionally, the websocket message handler will fail immediately if the server is suspended and a message is received.

Also blocks API requests from even establishing a websocket connection when the server is already suspended. Noticed a bug in the context cancelation behavior that was running all of the close logic, but never actually killing the loop so the connection just sat there open.

This is going to require some tweaks in the panel too — right now when you do this it just sits there flooding the `/ws` endpoint because it doesn't know how to handle this specific case.